Extract code to files
