+++ VoiceOver Designer/Features/Sources/Editor/FlippedView.swift
@@ -1,0 +1,13 @@
+import AppKit
+
+class FlippedView: NSView {
+    override var isFlipped: Bool {
+        true
+    }
+}
+
+class FlippedStackView: NSStackView {
+    override var isFlipped: Bool {
+        true
+    }
+}
